Add v2/scan endpoint
The new endpoint returns a json object which contains the response description from ClamAV.  This, combined with configuring ClamAV to block password protected files will allow application developers to determine if a file was blocked due to malware or because it is password protected and to response accordingly.

For legacy purposes the /scan endpoint remains untouched.